requirement.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
discord.py==1.7.3
discord.py-message-components==1.7.5.4
google-api-core==2.7.1
google-api-python-client==2.40.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
shortuuid==1.0.8
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
yarl==1.7.2
